% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstyperidge.R
\name{regstyperidge}
\alias{regstyperidge}
\title{Full regression results using the S-type robust ridge regression estimators}
\usage{
regstyperidge(x, y)
}
\arguments{
\item{x}{Explanatory variables (data.frame, matrix)}

\item{y}{Dependent variables (data.frame, vector)}
}
\value{
A list of lists
}
\description{
Full regression results using the S-type robust ridge regression estimators
}
\examples{
library("mctest")
x <- Hald[,-1]
y <- Hald[,1]
regstyperidge(x,y)

library(isdals)
data(bodyfat)
x <- bodyfat[,-1]
y <- bodyfat[,1]
regstyperidge(x,y)
}
